+++ src/Continuation.php
@@ -3,47 +3,42 @@ declare(strict_types = 1);
 
 namespace Innmind\IPC;
 
+use Innmind\IPC\Continuation\Next;
+use Innmind\Immutable\Sequence;
+
 /**
- * @template T
  * @psalm-immutable
+ * @template T
  */
 final class Continuation
 {
-    private Client $client;
-    /** @var T */
-    private mixed $carry;
-    private bool $closed;
-    private ?Message $response;
-    private bool $stop;
-
     /**
+     * @param Sequence<Message> $messages
      * @param T $carry
      */
     private function __construct(
-        Client $client,
-        mixed $carry,
-        bool $closed = false,
-        Message $response = null,
-        bool $stop = false,
+        private Next $next,
+        private Sequence $messages,
+        private mixed $carry,
     ) {
-        $this->client = $client;
-        $this->carry = $carry;
-        $this->closed = $closed;
-        $this->response = $response;
-        $this->stop = $stop;
     }
 
     /**
      * @internal
+     * @psalm-pure
      * @template A
      *
      * @param A $carry
      *
      * @return self<A>
      */
-    public static function start(Client $client, mixed $carry): self
+    public static function new(mixed $carry): self
     {
-        return new self($client, $carry);
+        return new self(
+            Next::continue,
+            Sequence::of(),
+            $carry,
+        );
     }
 
     /**
@@ -51,83 +46,62 @@ final class Continuation
      *
      * @return self<T>
      */
-    public function continue(mixed $carry): self
+    public function carryWith(mixed $carry): self
     {
-        return new self($this->client, $carry);
+        return new self(
+            $this->next,
+            $this->messages,
+            $carry,
+        );
     }
 
     /**
-     * This will send the given message to the client
-     *
-     * @param T $carry
+     * @param Message|Sequence<Message> $messages
      *
      * @return self<T>
      */
-    public function respond(mixed $carry, Message $message): self
+    public function respond(Message|Sequence $messages): self
     {
-        return new self($this->client, $carry, response: $message);
+        if ($messages instanceof Message) {
+            $messages = Sequence::of($messages);
+        }
+
+        return new self(
+            $this->next,
+            $messages->prepend($this->messages), // to keep the user provided lazyness
+            $this->carry,
+        );
     }
 
     /**
-     * The client will be closed and then garbage collected
-     *
-     * @param T $carry
-     *
      * @return self<T>
      */
-    public function close(mixed $carry): self
+    public function finish(): self
     {
-        return new self($this->client, $carry, closed: true);
-    }
-
-    /**
-     * The server will be gracefully shutdown
-     *
-     * @param T $carry
-     *
-     * @return self<T>
-     */
-    public function stop(mixed $carry): self
-    {
-        return new self($this->client, $carry, stop: true);
+        return new self(
+            Next::finish,
+            $this->messages,
+            $this->carry,
+        );
     }
 
     /**
      * @internal
-     * @template A
-     * @template B
-     * @template C
-     * @template D
+     * @template R
      *
-     * @param callable(Client, Message, T): A $onResponse
-     * @param callable(Client, T): B $onClose
-     * @param callable(Client, T): C $onStop
-     * @param callable(Client, T): D $onContinue
+     * @param callable(T, Sequence<Message>): R $continue
+     * @param callable(T, Sequence<Message>): R $finish
      *
-     * @return A|B|C|D
+     * @return R
      */
     public function match(
-        callable $onResponse,
-        callable $onClose,
-        callable $onStop,
-        callable $onContinue,
+        callable $continue,
+        callable $finish,
     ): mixed {
-        if ($this->response instanceof Message) {
-            /** @psalm-suppress ImpureFunctionCall */
-            return $onResponse($this->client, $this->response, $this->carry);
-        }
-
-        if ($this->closed) {
-            /** @psalm-suppress ImpureFunctionCall */
-            return $onClose($this->client, $this->carry);
-        }
-
-        if ($this->stop) {
-            /** @psalm-suppress ImpureFunctionCall */
-            return $onStop($this->client, $this->carry);
-        }
-
         /** @psalm-suppress ImpureFunctionCall */
-        return $onContinue($this->client, $this->carry);
+        return match ($this->next) {
+            Next::continue => $continue($this->carry, $this->messages),
+            Next::finish => $finish($this->carry, $this->messages),
+        };
     }
 }